backend/requirements.txt
# Django 核心（需 4.2.16+ 才官方支持 Python 3.13）
Django>=4.2.16,<4.3

# REST Framework
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2

# JWT Token - PyJWT 纯 Python 模式处理 HS256，不需要 cryptography
PyJWT==2.8.0

# 数据库驱动（生产环境可选）
# mysqlclient==2.2.4
# psycopg2-binary==2.9.9

# 图片处理（需 10.2+ 支持 Python 3.13）
Pillow>=10.2.0

# 生产服务器
gunicorn==21.2.0
whitenoise==6.6.0
waitress==2.1.2

# AI 服务
openai>=1.0.0

# 工具库
python-dotenv==1.0.1
